fix warnings

+++ cpp/gui_app/include/frontend_game.hh
@@ -23,10 +23,10 @@ namespace c20::gui {
 			std::pair<int, int> current_pos_coord;
 			std::array<std::array<std::string, M>, N> label_texts;
 
-			int height() const {return M;}
-			int width() const {return N;}
+			unsigned height() const {return M;}
+			unsigned width() const {return N;}
 			size_t size() const {return history.size();}
-			int current_idx() const {return current_pos_idx;}
+			unsigned current_idx() const {return current_pos_idx;}
 			const PositionWrapper& at(int idx) {return history[idx];}
 
 
@@ -44,7 +44,7 @@ namespace c20::gui {
 					current_pos_coord.first *M - current_pos_coord.second;
 				for (int i = 0; i < N; i++) {
 					for (int j = 0; j < M; j++) {
-						if (cur_idx < 0 || cur_idx >= history.size()) {
+						if (cur_idx < 0 || (unsigned) cur_idx >= history.size()) {
 							label_texts[i][j] = "";
 						} else {
 							label_texts[i][j] = fmt::format("{}", cur_idx);
